+++ visitor-main/src/main/java/org/visitor/appportal/web/controller/common/ProductController.java
@@ -256,14 +256,14 @@ public class ProductController extends BasicController {
 				
 				pmpT.setPmpProductId(product.getProductId());
 				pmpT.setPmpProductPriceKey(pdt.getAdditionalPriceKeyStr());
-				pmpT.setPmpProductPriceValue(Integer.valueOf(pdt.getAdditionalPriceValue()));
+				pmpT.setPmpProductPriceValue(Double.valueOf(pdt.getAdditionalPriceValue()));
 				pmpT.setPmpStatus(0);
 				
 				visitorProductMultiPriceService.saveProductMultiPrice(pmpT);
 				productRedisService.saveProductMultiPriceToRedis(pmpT);
 			} else {
 				ProductMultiPrice pmp = productRedisService.getProductMultiPriceSetByProductIdAndKey(product.getProductId(), pdt.getAdditionalPriceKeyStr());
-				pmp.setPmpProductPriceValue(Integer.valueOf(pdt.getAdditionalPriceValue()));
+				pmp.setPmpProductPriceValue(Double.valueOf(pdt.getAdditionalPriceValue()));
 				visitorProductMultiPriceService.saveProductMultiPrice(pmp);
 				productRedisService.saveProductMultiPriceToRedis(pmp);
 			}
